feat : step2
Calculator, Operator enum 클래스 분리
